initialized vite app

+++ backend/controller/user.controller.js
@@ -1,13 +1,25 @@
 import mongoose from 'mongoose';
 import User from "../models/user.model.js";
 
-export const getUser = async (req, res) => {
+export const getAllUsers = async (req, res) => {
     try {
         const users = await User.find({});
         res.status(200).json({ success: true, data: users});
     } catch (error)  {
         console.log("Error in fetching users:", error.message)
         res.status(500).json({ success: false, message: "Server Error" })
+    }
+};
+
+export const getUser = async (req, res) => {
+    const { id } = req.params;
+
+    try {
+        const user = await User.findById(id);
+        res.status(200).json({ success: true, data: user });
+    } catch (error) {
+        console.log("Error in fetching user:", error.message);
+        res.status(500).json({ success: false, message: "Server Error" });
     }
 };
 
@@ -46,12 +58,16 @@ export const updateUser = async (req, res) => {
 };
 
 export const deleteUser = async (req, res) => {
-    const {id} = req.params;
+    const { id } = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid User Id" });
+    }
+    
     try{
         await User.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "User Deleted" });
     } catch (error) {
-        res.status(404).json({ success: false, message: "User not found"});
+        res.status(500).json({ success: false, message: "Server Error"});
     }
 };
